cleaned up the main of the parser program
git-svn-id: https://repo.cs.iastate.edu/coms309/19/trunk@383 817968c2-256c-45f1-8ff6-e7619683c4bf

+++ Batch/Parser/ParserProgram.java
@@ -3,36 +3,28 @@ package Parser;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
-
 import libraryManager.Library;
 
+/**
+ * Main for the iTunesLibrary parser. Takes in a file path
+ * to the iTunes library xml file or uses the default path
+ * if one is not given. A copy of the file is created and
+ * then removed once the program is done with it. A library
+ * of tracks is generated from the file and then sent to 
+ * the database.
+ */
 public class ParserProgram 
 {
 	public static void main(String[] args) 
 	{
-		String path = "C:\\iTunes Library.xml";	
-		if(args.length == 1)
-		{
-			path = args[0];
-		}
-		
-	
+		String path = args.length == 1 ? args[0] : "C:\\iTunes Library.xml";		
 		File src = new File(path);
 		File directory = new File(src.getParent() + "DupLibrary");
 		File dest = new File(directory.getAbsolutePath() + "\\" + src.getName());
 		try 
 		{
-			if(!directory.exists())
-			{
-				directory.mkdir();
-			}
-			
-			
-			if(dest.exists())
-			{
-				dest.delete();
-			}
-			
+			if(!directory.exists()) { directory.mkdir(); }
+			if(dest.exists()) { dest.delete(); }
 			Files.copy(src.toPath(), dest.toPath());
 		} 
 		catch (IOException e) 
@@ -44,10 +36,7 @@ public class ParserProgram
 		library.createFromITunesDB(dest.getAbsolutePath());
     	library.addAllToDB();
     	
-    	if(directory.exists() && dest.exists())
-    	{
-    		System.out.println(dest.delete());
-    		System.out.println(directory.delete());
-    	}
+    	if(dest.exists()) { dest.delete(); }
+    	if(directory.exists()) { directory.delete(); }
 	}
 }
